% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Return dataframe from data file.}
\usage{
read_file(file)
}
\arguments{
\item{file}{A file, of any type.}
}
\value{
A dataframe of your data file.
}
\description{
Return dataframe from data file.
}
\examples{
read_file('cool.sas7bdat')
read_file('data.csv')
}
